Fixed circular dependency between FriendshipService and BlockedUserService
- BlockedUserService functions now only block and unblock users. The rest of the checks happen in the FriendshipService class.

+++ backend/src/main/java/com/fnafke/vexa/services/BlockedUserServiceImpl.java
@@ -1,17 +1,12 @@
 package com.fnafke.vexa.services;
 
-import java.util.UUID;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.fnafke.vexa.models.BlockedUser;
-import com.fnafke.vexa.models.Friendship;
 import com.fnafke.vexa.models.User;
 import com.fnafke.vexa.repositories.BlockedUserRepository;
 import com.fnafke.vexa.services.interfaces.BlockedUserService;
-import com.fnafke.vexa.services.interfaces.FriendshipService;
-import com.fnafke.vexa.services.interfaces.UserService;
 
 @Service
 public class BlockedUserServiceImpl implements BlockedUserService {
@@ -19,38 +14,21 @@ public class BlockedUserServiceImpl implements BlockedUserService {
     @Autowired
     private BlockedUserRepository blockedUserRepository;
 
-    @Autowired
-    private UserService userService;
-
-    @Autowired
-    private FriendshipService friendshipService;
-
     public BlockedUserServiceImpl(
-            BlockedUserRepository blockedUserRepository,
-            UserService userService,
-            FriendshipService friendshipService) {
+            BlockedUserRepository blockedUserRepository) {
         this.blockedUserRepository = blockedUserRepository;
-        this.userService = userService;
-        this.friendshipService = friendshipService;
     }
 
     @Override
-    public boolean isUserBlockedBy(UUID blockerId, UUID blockedId) {
-        return blockedUserRepository.existsByBlockerIdAndBlockedId(blockerId, blockedId);
+    public boolean isUserBlockedBy(User blocker, User blocked) {
+        return blockedUserRepository.existsByBlockerIdAndBlockedId(blocker.getId(), blocked.getId());
     }
 
     @Override
-    public BlockedUser blockUser(UUID blockerId, UUID blockedId) {
-        User blocker = userService.findById(blockerId);
-        User blocked = userService.findById(blockedId);
+    public BlockedUser blockUser(User blocker, User blocked) {
 
-        if (isUserBlockedBy(blockerId, blockedId)) {
+        if (isUserBlockedBy(blocker, blocked)) {
             throw new IllegalArgumentException("User is already blocked.");
-        }
-
-        Friendship friendship = friendshipService.getFriendshipBetweenUsers(blockerId, blockedId);
-        if (friendship != null) {
-            this.friendshipService.removeFriendship(friendship.getId());
         }
 
         BlockedUser blockedUser = new BlockedUser(blocker, blocked);
@@ -60,11 +38,12 @@ public class BlockedUserServiceImpl implements BlockedUserService {
     }
 
     @Override
-    public String unblockUser(UUID blockerId, UUID blockedId) {
-        User blocker = userService.findById(blockerId);
-        User blocked = userService.findById(blockedId);
+    public String unblockUser(User blocker, User blocked) {
+        if (!isUserBlockedBy(blocker, blocked)) {
+            throw new IllegalArgumentException("User is not blocked.");
+        }
 
-        BlockedUser blockedUser = blockedUserRepository.findBetweenUsers(blockerId, blockedId)
+        BlockedUser blockedUser = blockedUserRepository.findBetweenUsers(blocker.getId(), blocked.getId())
                 .orElseThrow(() -> new IllegalArgumentException("User is not blocked."));
 
         blockedUserRepository.delete(blockedUser);
